🔥 cleans up unused imports and marks some components that are out of commission temporarily (#831)
updates comment

adds comment
