+++ public/locales/fr/translations.json
@@ -95,9 +95,6 @@
   "- all files will be included in the manifest": "- all files will be included in the manifest",
   "- the participants in your query have related family member data": "- the participants in your query have related family member data",
   "To include the family data in the manifest, select your desired data types below": "To include the family data in the manifest, select your desired data types below",
-  "Generate manifest ID": "Generate manifest ID",
-  "Copy ID": "Copy ID",
-  "GENERATE": "GENERATE",
   "Download Manifest": "Download Manifest",
   "Settings": "Settings",
   "Integrations": "Integrations",
