+++ src/components/FamilyManifestModal/FamilyManifestModal.js
@@ -215,7 +215,6 @@ export default compose(
                             show: true,
                             type: 'string',
                             sortable: true,
-                            type: 'string',
                           },
                         ],
                       }),
