+++ src/components/FileRepo/AggregationSidebar/AggregationSidebar.js
@@ -13,7 +13,6 @@ import UploadIdsButton from './UploadIdsButton';
 import AdvancedFacetViewModalContent from 'components/AdvancedFacetViewModal';
 import { ScrollbarSize } from 'components/ContextProvider/ScrollbarSizeProvider';
 import { config as statsConfig } from 'components/Stats';
-import { ActionButton } from 'uikit/Button';
 import { TRACKING_EVENTS } from 'services/analyticsTracking';
 import { FilterInput } from 'uikit/Input';
 import Column from 'uikit/Column';
