+++ src/components/UserDashboard/MySavedQueries.js
@@ -2,9 +2,7 @@ import React, { Fragment } from 'react';
 import Spinner from 'react-spinkit';
 import styled from 'react-emotion';
 
-import TrashIcon from 'react-icons/lib/fa/trash';
 import { H2 } from 'uikit/Headings';
-import { H3 } from './styles';
 import { compose, lifecycle } from 'recompose';
 import { injectState } from 'freactal';
